third_party/WebKit/LayoutTests/imported/csswg-test/css-shapes-1/shape-outside/values/shape-image-threshold-002-expected.txt

This is a testharness.js-based test.
FAIL -0.5 - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL -100 - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL nonNumber - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL 3/4 - inline - invalid assert_equals: expected (object) null but got (string) ""
FAIL 20% - inline - invalid assert_equals: expected (object) null but got (string) ""
Harness: the test ran to completion.
